5week/실습6/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
Werkzeug==2.0.3
protobuf==3.19.4
MarkupSafe==2.0.1 
setuptools==40.8.0